[consensus-types] move accumulator_extension_proof to libra_types
